Add test for invalid author fields
