+++ src/definitions/song.ts
@@ -1,4 +1,5 @@
 import { LegacyCategory, Tag } from './tags';
+import { Author } from './author';
 
 export type Song = {
 	id: number;
@@ -10,14 +11,6 @@ export type Song = {
 	sorting?: number;
 	deleted?: true;
 	content: string;
-};
-
-export type Author = {
-	name?: string;
-	event?: string;
-	location?: string;
-	year?: number;
-	comment?: string;
 };
 
 export type XmlifyableSong = Song & {
